extracting methods to CrossTenantAssociations module
