+++ lib/active_record/tenanted/tenant.rb
@@ -85,6 +85,8 @@ module ActiveRecord
       CONNECTION_POOL_CREATION_LOCK = Thread::Mutex.new # :nodoc:
 
       class_methods do
+        include CrossTenantAssociations::ClassMethods
+
         def tenanted?
           true
         end
@@ -228,14 +230,6 @@ module ActiveRecord
           pool
         end
 
-        def has_one(name, scope = nil, **options)
-          p "HELLO"
-          p name
-          p scope
-          p options
-          enhanced_scope = enhance_cross_tenant_association(name, scope, options, :has_one)
-          super(name, enhanced_scope, **options)
-        end
 
         private
           def retrieve_connection_pool(strict:)
@@ -265,21 +259,6 @@ module ActiveRecord
             else
               yield
             end
-          end
-
-          def enhance_cross_tenant_association(name, scope, options, association_type)
-            target_class = options[:class_name]&.constantize || name.to_s.classify.constantize
-
-            unless target_class.tenanted?
-              tenant_column = options[:tenant_column] || :tenant_id
-
-              return ->(record) {
-                base_scope = scope ? target_class.instance_exec(&scope) : target_class.all
-                base_scope.where(tenant_column => record.tenant)
-              }
-            end
-
-            scope
           end
       end
